+++ R/Class-LAScatalog.R
@@ -84,7 +84,6 @@
 #' algorithm that requires a buffer. In rare cases it might be useful to disable these controls. If
 #' `wall.to.wall = FALSE` controls are disabled and wall-to-wall outputs cannot be guaranteed.
 #' See \link{opt_wall_to_wall}
-#' }
 #'
 #' @section Chunk options:
 #' The slot `@chunk_options` contains a `list` of options that determine how chunks
@@ -102,6 +101,11 @@
 #' chunk pattern. By default the alignment is made along (0,0), meaning that the edge of the first chunk
 #' will belong on x = 0 and y = 0 and all the the other chunks will be multiples of the chunk size.
 #' Not relevant if `chunk_size = 0`. See \link{opt_chunk_alignment}.
+#' - **drop**: integers. A vector of integers that specify the IDs of the chunks that should not be
+#' created. This is designed to enable to restart a computation that failed without reprocessing
+#' everything. See `opt_restart`. Technically this options may be used for partial processing of
+#' a collection but should not. Partial processing is already a feature of the engine. See
+#' [this vignette](https://cran.r-project.org/web/packages/lidR/vignettes/lidR-LAScatalog-engine.html#partial-processing)
 #'
 #' @section Output options:
 #' The slot `@output_options` contains a `list` of options that determine how chunks
@@ -248,7 +252,8 @@ setMethod("initialize", "LAScatalog", function(.Object)
   .Object@chunk_options <- list(
     size = 0,
     buffer = 30,
-    alignment = c(0,0)
+    alignment = c(0,0),
+    drop = NULL
   )
 
   .Object@processing_options <- list(